requirements.txt
tensorflow
keras
numpy
matplotlib
pandas
scikit-learn
opencv-python
pillow
seaborn
